pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "tenable-mcp-mssp"
version = "0.1.0"
description = "An MCP server for orchestrating Tenable MSSP child container workflows."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastmcp>=3.2.4",
    "pyTenable>=1.9.1",
]

[project.scripts]
tenable-mcp-mssp = "tenable_mcp_mssp.server:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["tenable_mcp_mssp*"]
